+++ Kitodo/src/main/java/de/sub/goobi/metadaten/FileManipulation.java
@@ -34,7 +34,7 @@ import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.kitodo.api.filemanagement.ProcessSubType;
 import org.kitodo.data.database.beans.Process;
-import org.kitodo.filters.FileDirectoryFilter;
+import org.kitodo.filters.IsDirectoryFilter;
 import org.kitodo.services.ServiceManager;
 import org.kitodo.services.file.FileService;
 
@@ -497,7 +497,7 @@ public class FileManipulation {
         this.allImportFolder = allImportFolder;
     }
 
-    private static FilenameFilter directoryFilter = new FileDirectoryFilter();
+    private static FilenameFilter directoryFilter = new IsDirectoryFilter();
 
     /**
      * Import files.
