+++ Kitodo/src/main/java/de/sub/goobi/persistence/apache/FolderInformation.java
@@ -26,9 +26,9 @@ import java.util.List;
 
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
-import org.kitodo.filters.FileBeginEndFilter;
-import org.kitodo.filters.FileEndFilter;
-import org.kitodo.filters.FileEndNotBeginFilter;
+import org.kitodo.filters.FileNameBeginsAndEndsWithFilter;
+import org.kitodo.filters.FileNameEndsWithFilter;
+import org.kitodo.filters.FileNameEndsAndDoesNotBeginWithFilter;
 import org.kitodo.services.ServiceManager;
 import org.kitodo.services.file.FileService;
 
@@ -60,7 +60,7 @@ public class FolderInformation {
         DIRECTORY_SUFFIX = ConfigCore.getParameter("DIRECTORY_SUFFIX", "tif");
         DIRECTORY_PREFIX = ConfigCore.getParameter("DIRECTORY_PREFIX", "orig");
         /* nur die _tif-Ordner anzeigen, die nicht mir orig_ anfangen */
-        FilenameFilter filterDirectory = new FileEndNotBeginFilter(DIRECTORY_PREFIX + "_", "_" + DIRECTORY_SUFFIX);
+        FilenameFilter filterDirectory = new FileNameEndsAndDoesNotBeginWithFilter(DIRECTORY_PREFIX + "_", "_" + DIRECTORY_SUFFIX);
         URI tifOrdner = null;
         ArrayList<URI> verzeichnisse = fileService.getSubUris(filterDirectory, dir);
 
@@ -139,7 +139,7 @@ public class FolderInformation {
             DIRECTORY_SUFFIX = ConfigCore.getParameter("DIRECTORY_SUFFIX", "tif");
             DIRECTORY_PREFIX = ConfigCore.getParameter("DIRECTORY_PREFIX", "orig");
             /* nur die _tif-Ordner anzeigen, die mit orig_ anfangen */
-            FilenameFilter filterDirectory = new FileBeginEndFilter(DIRECTORY_PREFIX + "_", "_" + DIRECTORY_SUFFIX);
+            FilenameFilter filterDirectory = new FileNameBeginsAndEndsWithFilter(DIRECTORY_PREFIX + "_", "_" + DIRECTORY_SUFFIX);
             URI origOrdner = null;
             ArrayList<URI> verzeichnisse = fileService.getSubUris(filterDirectory, dir);
             // TODO: does it actually make sense?
@@ -220,7 +220,7 @@ public class FolderInformation {
      */
     public URI getSourceDirectory() {
         URI dir = getImagesDirectory();
-        FilenameFilter filterDirectory = new FileEndFilter("_source");
+        FilenameFilter filterDirectory = new FileNameEndsWithFilter("_source");
         URI sourceFolder = null;
         ArrayList<URI> verzeichnisse = fileService.getSubUris(filterDirectory, dir);
         if (verzeichnisse == null || verzeichnisse.size() == 0) {
