+++ Kitodo/src/main/java/org/goobi/io/BackupFileRotation.java
@@ -20,7 +20,7 @@ import java.util.ArrayList;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.kitodo.data.database.beans.Process;
-import org.kitodo.filters.FileListFilter;
+import org.kitodo.filters.FileNameMatchesFilter;
 import org.kitodo.services.ServiceManager;
 import org.kitodo.services.file.FileService;
 
@@ -152,7 +152,7 @@ public class BackupFileRotation {
     private ArrayList<URI> generateBackupBaseNameFileList(String filterFormat, Process process) {
 
         ArrayList<URI> filteredUris = new ArrayList<>();
-        FilenameFilter filter = new FileListFilter(filterFormat);
+        FilenameFilter filter = new FileNameMatchesFilter(filterFormat);
 
         URI processDataDirectory = serviceManager.getProcessService().getProcessDataDirectory(process);
         ArrayList<URI> subUris = fileService.getSubUris(filter, processDataDirectory);
