+++ Kitodo/src/main/java/org/kitodo/filters/FileNameEndsWithFilter.java
@@ -14,27 +14,30 @@ package org.kitodo.filters;
 import java.io.File;
 import java.io.FilenameFilter;
 
-public class FileListFilter implements FilenameFilter {
+/**
+ * Implementation of FileNameFilter for checking the end of files.
+ */
+public class FileNameEndsWithFilter implements FilenameFilter {
 
-    private String name;
+    private String end;
 
     /**
-     * Filter file list according to the given name.
+     * Filter files by given file end.
      *
-     * @param name
-     *            String
+     * @param end
+     *            file name should end with this String
      * @throws IllegalArgumentException
      *             it is thrown in case parameter is null or empty String
      */
-    public FileListFilter(String name) {
-        if (name == null || name.equals("")) {
-            throw new IllegalArgumentException("No filter or empty name is given.");
+    public FileNameEndsWithFilter(String end) {
+        if (end == null || end.equals("")) {
+            throw new IllegalArgumentException("No filter or empty filter for file end is given.");
         }
-        this.name = name;
+        this.end = end;
     }
 
     @Override
-    public boolean accept(File directory, String filename) {
-        return filename.matches(name);
+    public boolean accept(File dir, String name) {
+        return (name.endsWith(end));
     }
 }
